Kafka: Offset
- Implemented Kafka producer and consumer for Baggage.
- Producer generates and sends random baggage events.
- Consumer logs baggage details along with offset information.

In Kafka, an offset is a sequential ID that uniquely identifies a record within a partition. This concept is similar to auto-increment fields in databases. Each message written to a partition gets a unique offset number.
Consumers keep track of the last offset they have read, which allows them to resume from the correct position.This ensures messages are not skipped or read multiple times.Offset management is handled through consumer groups, which coordinate how messages are consumed.Consumers with the same groupId share the message load, while consumers with different groupIds read all messages independently. This can be observed when the application and Kafka are restarted and a request is made again.