Kafka: Retry Mechanism & Error Handling
In this example, a ticket purchase simulation with a retry mechanism was performed.

The first event falls into the "ticket-purchase-requests topic". If it cannot be processed, it is thrown into the -> "ticket-purchase-retry topic" (max 3 attempts). If the retry attempts also fail, it falls into the -> "ticket-purchase-dlt" topic. You can observe these topics with kafka-ui.