Kafka: Transactional Outbox Design Pattern
Implementing the Outbox Pattern using Spring Boot, Kafka, and JPA to ensure reliable and consistent event publishing.

By sending multiple transfer requests, you'll see that the "processed" field of the transfers sent is "false". Then, when you see that the message is consumed by the consumer, you'll notice this field is set to "true". Since both database records occur within the same transaction, either both succeed or both fail. This way, unprocessed messages are re-processed, and outbox design pattern applications are implemented. Also, note that the object itself is stored in the "payload" section. Other fields can be changed according to need. Lastly, "aggregate_id" is a foreign key that indicates which soccer player this event is associated with.